removed logger, now logging happens directly from the runner, therefore now the runner interface expose functions for logging
