+++ zenoh-flow/src/runtime/dataflow/instance/runners/sink.rs
@@ -108,6 +108,18 @@ impl Runner for SinkRunner {
         inputs
     }
 
+    async fn start_recording(&self) -> ZFResult<String> {
+        Err(ZFError::Unimplemented)
+    }
+
+    async fn stop_recording(&self) -> ZFResult<String> {
+        Err(ZFError::Unimplemented)
+    }
+
+    async fn is_recording(&self) -> bool {
+        false
+    }
+
     async fn run(&self) -> ZFResult<()> {
         let mut context = Context::default();
